add batch list item component and change kibana link
